feat: separate 404 not found page ui and business logic
#41
